+++ package.json
@@ -8,11 +8,11 @@
     "bootstrap": "^4.2.1",
     "express": "^4.17.3",
     "react": "^16.7.0",
-    "react-chat-widget": "^2.1.5",
+    "react-chat-widget": "^2.1.4-1",
     "react-dom": "^16.7.0",
     "react-md-spinner": "^0.3.0",
-    "react-router-dom": "^0.0.0-experimental-compat.6",
-    "react-scripts": "^2.1.3"
+    "react-router-dom": "^4.3.1",
+    "react-scripts": "2.1.3"
   },
   "scripts": {
     "start": "react-scripts start",
