+++ src/Client.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import {Widget, addResponseMessage, addUserMessage, dropMessages} from 'react-chat-widget';
 import {CometChat} from '@cometchat-pro/chat';
 
-import config from '../config';
+import config from './config';
 import 'react-chat-widget/lib/styles.css';
 
 const agentUID = config.agentUID;
@@ -156,8 +156,7 @@ class Client extends Component {
   render() {
     return (
       <div className='App'>
-        <Widget 
-
+        <Widget
           handleNewUserMessage={this.handleNewUserMessage}
           title='My E-commerce Live Chat'
           subtitle='Ready to help you'
