--- src/components/CSS/Authentification.css
@@ -1,5 +1,0 @@
-.main {
-    background-image: url("books.jpg");
-    background-size: 1500px;
-    background-position:center;
-}
