--- src/components/CSS/Home.css
@@ -1,34 +1,0 @@
-.under {
-    position: absolute; 
-    top: 0; 
-    right: 0; 
-    bottom: 0; 
-    left: 0;
-    background-image: url("bookshelves.jpg") !important;
-    background-size: 1500px;
-    background-position:center;
-}
-
-.over {
-    position: absolute; 
-    top: 0; 
-    right: 0; 
-    bottom: 0; 
-    left: 0;
-    background: rgba(60, 52, 36, 0.644);
-    color: white;
-}
-
-h1 {
-    padding-top: 200px;
-    padding-bottom:50px;
-}
-
-.conversation-container > .header {
-    background-color: red;
-}
-
-.message > .response {
-    background-color: black;
-    color: white;
-}
