+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.css';
 import './index.css';
-import App from './components/App';
+import App from './App';
 import { CometChat } from '@cometchat-pro/chat';
 import config from './config';
